+++ src/primitives/Hash.ts
@@ -168,48 +168,42 @@ abstract class BaseHash {
    * @returns Returns an array denoting the padding.
    */
   private _pad (): number[] {
-    //
-    let len = this.pendingTotal
+    const len = this.pendingTotal
     const bytes = this._delta8
     const k = bytes - ((len + this.padLength) % bytes)
     const res = new Array(k + this.padLength)
     res[0] = 0x80
-    let i
+    let i: number
     for (i = 1; i < k; i++) {
       res[i] = 0
     }
 
     // Append length
-    len <<= 3
-    let t
-    if (this.endian === 'big') {
-      for (t = 8; t < this.padLength; t++) {
-        res[i++] = 0
-      }
+    const lengthBytes = this.padLength
+    const maxBits = 1n << BigInt(lengthBytes * 8)
+    let totalBits = BigInt(len) * 8n
 
-      res[i++] = 0
-      res[i++] = 0
-      res[i++] = 0
-      res[i++] = 0
-      res[i++] = (len >>> 24) & 0xff
-      res[i++] = (len >>> 16) & 0xff
-      res[i++] = (len >>> 8) & 0xff
-      res[i++] = len & 0xff
-    } else {
-      res[i++] = len & 0xff
-      res[i++] = (len >>> 8) & 0xff
-      res[i++] = (len >>> 16) & 0xff
-      res[i++] = (len >>> 24) & 0xff
-      res[i++] = 0
-      res[i++] = 0
-      res[i++] = 0
-      res[i++] = 0
-
-      for (t = 8; t < this.padLength; t++) {
-        res[i++] = 0
-      }
+    if (totalBits >= maxBits) {
+      throw new Error('Message too long for this hash function')
     }
 
+    if (this.endian === 'big') {
+      const lenArray = new Array<number>(lengthBytes)
+
+      for (let b = lengthBytes - 1; b >= 0; b--) {
+        lenArray[b] = Number(totalBits & 0xffn)
+        totalBits >>= 8n
+      }
+
+      for (let b = 0; b < lengthBytes; b++) {
+        res[i++] = lenArray[b]
+      }
+    } else {
+      for (let b = 0; b < lengthBytes; b++) {
+        res[i++] = Number(totalBits & 0xffn)
+        totalBits >>= 8n
+      }
+    }
     return res
   }
 }